renamed folder of Modal Component
